requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-genai
pillow
yt-dlp
openai-whisper
torch
transformers
requests